# storage/api-docs/api-docs-admin-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation for admin",
        "description": "L5 Swagger OpenApi description for admin",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/admin/v1/store-activity": {
            "post": {
                "tags": [
                    "Activity"
                ],
                "operationId": "createActivity",
                "requestBody": {
                    "description": "Body request needed to add activity object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "name"
                                    },
                                    "price": {
                                        "description": "price",
                                        "type": "number",
                                        "format": "double"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKey": []
                    }
                ]
            }
        },
        "/admin/v1/get-activities": {
            "get": {
                "tags": [
                    "Activity"
                ],
                "summary": "Rertrieve all activities",
                "operationId": "getAllActivities",
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "APIKey": []
                    }
                ]
            }
        },
        "/admin/v1/get-activity-by-slug/{activity_slug}": {
            "get": {
                "tags": [
                    "Activity"
                ],
                "summary": "Retrieve specific activity",
                "operationId": "getActivityBySlug",
                "parameters": [
                    {
                        "name": "activity_slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "APIKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "APIKey": {
                "type": "apiKey",
                "name": "X-Authorization",
                "in": "header"
            }
        }
    }
}